add last minute buffer; fix multiple bugs
Added a command to fetch the values of a given variable in the last minute.
Fixed timer bug in TCP server that would allow the callback function to be called after the TCP session object was destroyed.
Added the write to stream functions.
Should probably change timestamps to chrono objects.